+++ src/collective/cover/browser/contentchooser.py
@@ -20,17 +20,16 @@ from zope.interface import Interface
 from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleTerm
 
-try:
-    import json
-    json = json  # Pyflakes
-except ImportError:
-    import simplejson as json
+import json
 
 VOCAB_ID = u'plone.app.vocabularies.ReallyUserFriendlyTypes'
 
 grok.templatedir("contentchooser_templates")
 
 
+# XXX: what's the purpose of this view?
+#      why is here if it's intendes for tests?
+#      can we get rid of it?
 class TestContent(grok.View):
     """
     test contentchooser for selecting