+++ src/collective/cover/controlpanel.py
@@ -1,12 +1,11 @@
 # -*- coding: utf-8 -*-
 
+from collective.cover import _
+from collective.cover.config import DEFAULT_AVAILABLE_TILES
+from collective.cover.config import DEFAULT_SEARCHABLE_CONTENT_TYPES
+from plone.app.registry.browser import controlpanel
 from zope import schema
 from zope.interface import Interface
-
-from plone.app.registry.browser import controlpanel
-
-from collective.cover import _
-from collective.cover.config import DEFAULT_SEARCHABLE_CONTENT_TYPES
 
 
 class ICoverSettings(Interface):
@@ -18,7 +17,17 @@ class ICoverSettings(Interface):
         required=True,
         key_type=schema.TextLine(title=_(u'Name')),
         value_type=schema.TextLine(title=_(u'Layout')),
-        readonly=True,)  # FIXME: we have no widget for this field yet
+        readonly=True,  # FIXME: we have no widget for this field yet
+    )
+
+    available_tiles = schema.List(
+        title=_(u"Available tiles"),
+        description=_(u"This tiles will be available for layout creation."),
+        required=True,
+        default=DEFAULT_AVAILABLE_TILES,
+        value_type=schema.Choice(
+            vocabulary=u'collective.cover.EnabledTiles'),
+    )
 
     searchable_content_types = schema.List(
         title=_(u"Searchable Content Types"),
@@ -28,7 +37,8 @@ class ICoverSettings(Interface):
         default=DEFAULT_SEARCHABLE_CONTENT_TYPES,
         # we are going to list only the main content types in the widget
         value_type=schema.Choice(
-            vocabulary=u'collective.cover.AvailableContentTypes'),)
+            vocabulary=u'collective.cover.AvailableContentTypes'),
+    )
 
 
 class CoverSettingsEditForm(controlpanel.RegistryEditForm):
@@ -42,6 +52,7 @@ class CoverSettingsEditForm(controlpanel.RegistryEditForm):
 
     def updateWidgets(self):
         super(CoverSettingsEditForm, self).updateWidgets()
+        self.widgets['available_tiles'].style = u'min-width: 200px;'
         self.widgets['searchable_content_types'].style = u'min-width: 200px;'